requirements.txt
psycopg2-binary
sqlalchemy
flask
python-dotenv
pydantic